link directly to the developer setup guide
Currently, in the Server and Web App "Workflow" guides, there is copy to "set up your developer environment" which links to https://docs.mattermost.com/developer/dev-setup.html, which subsequently links the user back to the previous guide in the respective track ("Developer Setup"). This changes the markdown to link directly to the previous "Developer Setup" guide.