Adjust formatting for "Install dependencies" (#532)
* Adjust formatting for "Install dependencies"

While running through installation it took a moment to understand that "Install dependencies" wasn't a separate instruction from the bullet points below it. Adding an indentation to the bullet points and a colon after "Install dependencies" makes this more clear imo.

* Updated run app instructions

After searching through GitHub issues and the Mattermost Developer channel I came across someone mentioning using 'make run' from the webapp directory to run the app. It worked for me and is missing from the docs (if this is indeed the correct approach and not something that just worked in my case: OSX 10.13.4)