Remove underscore from action id
According to https://mattermost.atlassian.net/browse/MM-49350, underscores are not allowed in action ids. I stumbled upon that some time ago and a [user in the forums](https://forum.mattermost.com/t/interactive-message-menu-not-returning-value/10841/10) brought this issue back to my attention, so here's a fix for the documentation, at least to make sure the docs have a hint about that and users (like me) don't copy a broken example config.