Update links to mattermost-redux source code after migration
The current docs point to
https://github.com/mattermost/mattermost-redux, which has migrated to
https://github.com/mattermost/mattermost-webapp, which has itself
migrated to https://github.com/mattermost/mattermost. So it's not very
convienient to access the current version of the source code.

This PR updates those links accordingly, so that the reader doesn't have
to jump through so many hoops.

I removed the line numbers for links which pointed to the master branch
instead of a specific commit as they get outdated as the source code
changes. The reader can easily find the function using GitHub's menu
or via text search. One could also convert those to permalinks (linking
to the current commit), which has itself downsides as the reader will
eventually be pointed to outdated source code.