Update e2e-cheatsheets.md (#1446)
* Update e2e-cheatsheets.md

The original link leads to Chrome extension that is no longer unavailable, so I've updated the docs to point to the always-maintained, zero-install online Testing Playground and the VS Code Testing Playground extension—ensuring users get up-to-date query guidance in both browser and editor environments. Hopefully, that's a suitable fix for you! :)

* Update site/content/contribute/more-info/webapp/e2e-cheatsheets.md

---------

Co-authored-by: Carrie Warner (Mattermost) <74422101+cwarnermm@users.noreply.github.com>